requirements.txt
uvicorn~=0.24.0
fastapi~=0.104.1
pydantic~=2.10.0
SQLAlchemy~=2.0.44
psycopg[binary]==3.2.3
python-dotenv~=1.0.0
gunicorn~=21.2.0